extract pdf code to sepparate feature module
