+++ core/testing/build.gradle.kts
@@ -13,11 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-plugins { alias(libs.plugins.kotlin.multiplatform) }
+plugins {
+  alias(libs.plugins.kotlin.multiplatform)
+  alias(libs.plugins.android.kotlin.multiplatform.library)
+  alias(libs.plugins.android.lint)
+}
 
 kotlin {
-  jvm()
   jvmToolchain(libs.versions.jvmToolchainVersion.get().toInt())
+  androidLibrary {
+    namespace = "org.groundplatform.core.testing"
+    compileSdk = libs.versions.androidCompileSdk.get().toInt()
+    minSdk = libs.versions.androidMinSdk.get().toInt()
+  }
+
+  jvm()
 
   iosArm64()
   iosSimulatorArm64()
@@ -26,10 +36,12 @@ kotlin {
     commonMain {
       dependencies {
         implementation(project(":core:domain"))
+        implementation(project(":core:ui"))
         implementation(libs.kotlin.stdlib)
         implementation(libs.kotlinx.collections.immutable)
         implementation(libs.kotlinx.coroutines.core)
         implementation(libs.kotlinx.serialization.json)
+        implementation(libs.compose.components.resources)
       }
     }
   }
