+++ core/testing/src/commonMain/kotlin/org/groundplatform/testing/FakeDateFormatter.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Google LLC
+ * Copyright 2026 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.groundplatform.testing
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        google()
-        mavenCentral()
-    }
+import org.groundplatform.ui.util.DateFormatter
+
+/** [DateFormatter] for tests, so assertions don't depend on the host locale or time zone. */
+object FakeDateFormatter : DateFormatter {
+
+  override fun formatDate(millis: Long): String = "DATE($millis)"
+
+  override fun formatTime(millis: Long): String = "TIME($millis)"
 }
-
-plugins {
-    id 'org.gradle.toolchains.foojay-resolver-convention' version '1.0.0'
-}
-
-include ':app', ':e2eTest'
-include ':core:ui'
-include ':core:domain'
-include ':core:testing'
-include ':feature:pdf'
