--- core/ui/src/androidMain/kotlin/org/groundplatform/ui/system/pdf/render/image/PdfImage.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2026 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.groundplatform.ui.system.pdf.render.image
-
-import android.graphics.Bitmap
-
-/**
- * Android wraps a [android.graphics.Bitmap]. The renderer reads [bitmap] directly to draw on a
- * Canvas.
- */
-actual data class PdfImage(val bitmap: Bitmap) {
-  actual val width: Int
-    get() = bitmap.width
-
-  actual val height: Int
-    get() = bitmap.height
-}
