--- core/ui/src/commonMain/kotlin/org/groundplatform/ui/system/pdf/PdfImageProvider.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright 2026 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.groundplatform.ui.system.pdf
-
-import org.groundplatform.ui.system.pdf.render.image.PdfImageSet
-
-/**
- * Platform abstraction for images (QR code and photos) needed for PDF rendering. Implementations
- * should handle bitmap decoding and resource lifecycle management.
- */
-interface PdfImageProvider {
-  suspend fun load(qrContent: String?, photoFilenames: Set<String>): PdfImageSet
-}
