--- core/ui/src/commonMain/kotlin/org/groundplatform/ui/system/pdf/PdfReportLauncher.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright 2026 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.groundplatform.ui.system.pdf
-
-/**
- * Presents the two terminal actions on a generated report: share via system share sheet, open in an
- * external viewer.
- */
-interface PdfReportLauncher {
-  fun share(path: String)
-
-  fun open(path: String)
-}
