+++ feature/pdf/src/commonMain/kotlin/org/groundplatform/feature/pdf/PdfRenderer.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Google LLC
+ * Copyright 2026 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.groundplatform.feature.pdf
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        google()
-        mavenCentral()
-    }
+import org.groundplatform.feature.pdf.model.SubmissionPdfDocument
+import org.groundplatform.feature.pdf.render.image.PdfImageSet
+
+/**
+ * Rasterises a [SubmissionPdfDocument] to a PDF file. Each platform should use its native text
+ * layout and PDF APIs to handle wrapping, pagination, and drawing. Writes the result to the
+ * provided output path.
+ */
+interface PdfRenderer {
+  suspend fun render(document: SubmissionPdfDocument, images: PdfImageSet, outputPath: String)
 }
-
-plugins {
-    id 'org.gradle.toolchains.foojay-resolver-convention' version '1.0.0'
-}
-
-include ':app', ':e2eTest'
-include ':core:ui'
-include ':core:domain'
-include ':core:testing'
-include ':feature:pdf'
