+++ feature/pdf/src/commonMain/kotlin/org/groundplatform/feature/pdf/render/image/PdfImage.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Google LLC
+ * Copyright 2026 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.groundplatform.feature.pdf.render.image
 
-pluginManagement {
-    repositories {
-        gradlePluginPortal()
-        google()
-        mavenCentral()
-    }
+/**
+ * Opaque platform image handle (Android `Bitmap`, iOS `CGImageRef`). The renderer reads the
+ * platform value via the `actual` declaration.
+ */
+expect class PdfImage {
+  val width: Int
+  val height: Int
 }
-
-plugins {
-    id 'org.gradle.toolchains.foojay-resolver-convention' version '1.0.0'
-}
-
-include ':app', ':e2eTest'
-include ':core:ui'
-include ':core:domain'
-include ':core:testing'
-include ':feature:pdf'
