+++ feature/pdf/src/commonMain/kotlin/org/groundplatform/feature/pdf/render/image/PdfImageSet.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.groundplatform.feature.pdf.render.image
+
+/**
+ * The set of images needed to render a single report. Owns the lifecycle of the platform handles,
+ * call [release] in a `finally` block to free native resources.
+ */
+class PdfImageSet(
+  private val images: Map<ImageRef, PdfImage>,
+  private val onRelease: () -> Unit = {},
+) {
+  operator fun get(ref: ImageRef): PdfImage? = images[ref]
+
+  fun release() {
+    onRelease()
+  }
+
+  sealed interface ImageRef {
+    data object Qr : ImageRef
+
+    data class Photo(val filename: String) : ImageRef
+  }
+}
